+++ src/operations.ts
@@ -126,6 +126,7 @@ export const OPENCV_OPERATIONS = [
   { method: "ellipse2Poly", wasmExport: "ellipse2Poly" },
   { method: "exp", wasmExport: "matExpInto" },
   { method: "flip", wasmExport: "matFlip" },
+  { method: "findContours", wasmExport: "matFindContoursInto" },
   { method: "cvtColor", wasmExport: "matCvtColorInto" },
   { method: "getAffineTransform", wasmExport: "matGetAffineTransform" },
   { method: "getLogLevel", wasmExport: "getLogLevel" },